+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix bridge bug.
 - Fix weather spawning at Lara position instead of camera position.
 - Fix lara collision with objects on water surface.
+- Fix darts scale.
 
 ### tomb3 additions